add: move user credentials to resource model
